project_06/py_cloud_functions/ip_locations/requirements.txt
functions-framework>=3.0.0,<4.0.0
google-cloud-bigquery>=3.0.0,<4.0.0
google-cloud-storage>=2.0.0,<3.0.0
ijson>=3.0.0